+++ axum/src/body/mod.rs
@@ -22,7 +22,17 @@ pub use bytes::Bytes;
 pub type BoxBody = http_body::combinators::UnsyncBoxBody<Bytes, Error>;
 
 /// Convert a [`http_body::Body`] into a [`BoxBody`].
+#[deprecated(note = "use `axum::body::boxed`", since = "0.3.4")]
 pub fn box_body<B>(body: B) -> BoxBody
+where
+    B: http_body::Body<Data = Bytes> + Send + 'static,
+    B::Error: Into<BoxError>,
+{
+    boxed(body)
+}
+
+/// Convert a [`http_body::Body`] into a [`BoxBody`].
+pub fn boxed<B>(body: B) -> BoxBody
 where
     B: http_body::Body<Data = Bytes> + Send + 'static,
     B::Error: Into<BoxError>,
@@ -31,5 +41,5 @@ where
 }
 
 pub(crate) fn empty() -> BoxBody {
-    box_body(http_body::Empty::new())
+    boxed(http_body::Empty::new())
 }